move cuda_setup file to install directory
