+++ install/cuda_setup.py
